@@ -12,6 +12,7 @@
   "unpkg": "d3-composite-projections.min.js",
   "module": "./index.js",
   "jsnext:main": "index",
+  "type": "module",
   "repository": {
     "type": "git",
     "url": "https://github.com/rveciana/d3-composite-projections.git"
@@ -24,17 +25,17 @@
     "postpublish": "git push && git push --tags"
   },
   "dependencies": {
-    "d3-geo": "^2.0.1",
-    "d3-path": "^2.0.0"
+    "d3-geo": "^3.0.1",
+    "d3-path": "^3.0.1"
   },
   "devDependencies": {
-    "canvas": "2.6.1",
-    "eslint": "^7",
-    "pixelmatch": "^5.2.1",
-    "pngjs": "^5.0.0",
-    "rollup": "^2.28.2",
-    "rollup-plugin-terser": "^7.0.2",
-    "tape": "^5.0.1",
-    "topojson": "3"
+    "@rollup/plugin-terser": "^0.1.0",
+    "canvas": "2.10.2",
+    "eslint": "^8.29.0",
+    "pixelmatch": "^5.3.0",
+    "pngjs": "^6.0.0",
+    "rollup": "^3.5.1",
+    "tape": "^5.6.1",
+    "topojson-client": "^3.1.0"
   }
 }
