@@ -1,21 +1,26 @@
-var tape = require("tape");
+import tape from "tape";
 
-tape.Test.prototype.inDelta = function(actual, expected, delta) {
+tape.Test.prototype.inDelta = function (actual, expected, delta) {
   delta = delta || 1e-6;
   this._assert(inDelta(actual, expected, delta), {
     message: "should be in delta " + delta,
     operator: "inDelta",
     actual: actual,
-    expected: expected
+    expected: expected,
   });
 };
 
-function inDelta(actual, expected, delta) {
-  return (Array.isArray(expected) ? inDeltaArray : inDeltaNumber)(actual, expected, delta);
+export function inDelta(actual, expected, delta) {
+  return (Array.isArray(expected) ? inDeltaArray : inDeltaNumber)(
+    actual,
+    expected,
+    delta
+  );
 }
 
 function inDeltaArray(actual, expected, delta) {
-  var n = expected.length, i = -1;
+  var n = expected.length,
+    i = -1;
   if (actual.length !== n) return false;
   while (++i < n) if (!inDelta(actual[i], expected[i], delta)) return false;
   return true;
