@@ -1,9 +1,12 @@
-import { terser } from "rollup-plugin-terser";
-import * as meta from "./package.json";
+import terser from "@rollup/plugin-terser";
+
+import { readFileSync } from "fs";
+
+const meta = JSON.parse(readFileSync("./package.json"), "utf8");
 
 const config = {
   input: "index.js",
-  external: Object.keys(meta.dependencies || {}).filter(key =>
+  external: Object.keys(meta.dependencies || {}).filter((key) =>
     /^d3-/.test(key)
   ),
   output: {
@@ -18,11 +21,11 @@ const config = {
     globals: Object.assign(
       {},
       ...Object.keys(meta.dependencies || {})
-        .filter(key => /^d3-/.test(key))
-        .map(key => ({ [key]: "d3" }))
-    )
+        .filter((key) => /^d3-/.test(key))
+        .map((key) => ({ [key]: "d3" }))
+    ),
   },
-  plugins: []
+  plugins: [],
 };
 
 export default [
@@ -31,15 +34,15 @@ export default [
     ...config,
     output: {
       ...config.output,
-      file: `${meta.name}.min.js`
+      file: `${meta.name}.min.js`,
     },
     plugins: [
       ...config.plugins,
       terser({
         output: {
-          preamble: config.output.banner
-        }
-      })
-    ]
-  }
+          preamble: config.output.banner,
+        },
+      }),
+    ],
+  },
 ];
