@@ -1,45 +1,42 @@
-import { epsilon } from "./math";
+import { epsilon } from "./math.js";
 import { geoAlbers as albers } from "d3-geo";
-import { fitExtent, fitSize } from "./fit";
+import { fitExtent, fitSize } from "./fit.js";
 import { path } from "d3-path";
 
 function multiplex(streams) {
   var n = streams.length;
   return {
-    point: function(x, y) {
+    point: function (x, y) {
       var i = -1;
       while (++i < n) streams[i].point(x, y);
     },
-    sphere: function() {
+    sphere: function () {
       var i = -1;
       while (++i < n) streams[i].sphere();
     },
-    lineStart: function() {
+    lineStart: function () {
       var i = -1;
       while (++i < n) streams[i].lineStart();
     },
-    lineEnd: function() {
+    lineEnd: function () {
       var i = -1;
       while (++i < n) streams[i].lineEnd();
     },
-    polygonStart: function() {
+    polygonStart: function () {
       var i = -1;
       while (++i < n) streams[i].polygonStart();
     },
-    polygonEnd: function() {
+    polygonEnd: function () {
       var i = -1;
       while (++i < n) streams[i].polygonEnd();
-    }
+    },
   };
 }
 
-export default function() {
+export default function () {
   var cache,
     cacheStream,
-    main = albers()
-      .rotate([4.4, 0.8])
-      .center([0, 55.4])
-      .parallels([50, 60]),
+    main = albers().rotate([4.4, 0.8]).center([0, 55.4]).parallels([50, 60]),
     mainPoint,
     shetland = albers()
       .rotate([4.4, 0.8])
@@ -48,14 +45,14 @@ export default function() {
     shetlandPoint,
     point,
     pointStream = {
-      point: function(x, y) {
+      point: function (x, y) {
         point = [x, y];
-      }
+      },
     };
 
   var shetlandBbox = [
     [-2.1, 70],
-    [-0.7, 59.8]
+    [-0.7, 59.8],
   ];
   function albersUk(coordinates) {
     var x = coordinates[0],
@@ -66,40 +63,39 @@ export default function() {
     );
   }
 
-  albersUk.invert = function(coordinates) {
+  albersUk.invert = function (coordinates) {
     var k = main.scale(),
       t = main.translate(),
       x = (coordinates[0] - t[0]) / k,
       y = (coordinates[1] - t[1]) / k;
 
-    return (y >= -0.089 && y < 0.06 && x >= 0.029 && x < 0.046
-      ? shetland
-      : main
+    return (
+      y >= -0.089 && y < 0.06 && x >= 0.029 && x < 0.046 ? shetland : main
     ).invert(coordinates);
   };
 
-  albersUk.stream = function(stream) {
+  albersUk.stream = function (stream) {
     return cache && cacheStream === stream
       ? cache
       : (cache = multiplex([
           main.stream((cacheStream = stream)),
-          shetland.stream(stream)
+          shetland.stream(stream),
         ]));
   };
 
-  albersUk.precision = function(_) {
+  albersUk.precision = function (_) {
     if (!arguments.length) return main.precision();
     main.precision(_), shetland.precision(_);
     return reset();
   };
 
-  albersUk.scale = function(_) {
+  albersUk.scale = function (_) {
     if (!arguments.length) return main.scale();
     main.scale(_), shetland.scale(_);
     return albersUk.translate(main.translate());
   };
 
-  albersUk.translate = function(_) {
+  albersUk.translate = function (_) {
     if (!arguments.length) return main.translate();
     var k = main.scale(),
       x = +_[0],
@@ -109,7 +105,7 @@ export default function() {
       .translate(_)
       .clipExtent([
         [x - 0.065 * k, y - 0.089 * k],
-        [x + 0.075 * k, y + 0.089 * k]
+        [x + 0.075 * k, y + 0.089 * k],
       ])
       .stream(pointStream);
 
@@ -117,18 +113,18 @@ export default function() {
       .translate([x + 0.01 * k, y + 0.025 * k])
       .clipExtent([
         [x + 0.029 * k + epsilon, y - 0.089 * k + epsilon],
-        [x + 0.046 * k - epsilon, y - 0.06 * k - epsilon]
+        [x + 0.046 * k - epsilon, y - 0.06 * k - epsilon],
       ])
       .stream(pointStream);
 
     return reset();
   };
 
-  albersUk.fitExtent = function(extent, object) {
+  albersUk.fitExtent = function (extent, object) {
     return fitExtent(albersUk, extent, object);
   };
 
-  albersUk.fitSize = function(size, object) {
+  albersUk.fitSize = function (size, object) {
     return fitSize(albersUk, size, object);
   };
 
@@ -137,7 +133,7 @@ export default function() {
     return albersUk;
   }
 
-  albersUk.drawCompositionBorders = function(context) {
+  albersUk.drawCompositionBorders = function (context) {
     /*var ul = main.invert([
       shetland.clipExtent()[0][0],
       shetland.clipExtent()[0][1]
@@ -177,7 +173,7 @@ export default function() {
     context.lineTo(ll[0], ll[1]);
     context.closePath();
   };
-  albersUk.getCompositionBorders = function() {
+  albersUk.getCompositionBorders = function () {
     var context = path();
     this.drawCompositionBorders(context);
     return context.toString();
