@@ -1,51 +1,87 @@
-import {epsilon} from "./math";
-import {geoMercator as mercator} from "d3-geo";
-import {fitExtent, fitSize} from "./fit";
-import {path} from "d3-path";
-
+import { epsilon } from "./math.js";
+import { geoMercator as mercator } from "d3-geo";
+import { fitExtent, fitSize } from "./fit.js";
+import { path } from "d3-path";
 
 // The projections must have mutually exclusive clip regions on the sphere,
 // as this will avoid emitting interleaving lines and polygons.
 function multiplex(streams) {
   var n = streams.length;
   return {
-    point: function(x, y) { var i = -1; while (++i < n) {streams[i].point(x, y); }},
-    sphere: function() { var i = -1; while (++i < n) {streams[i].sphere(); }},
-    lineStart: function() { var i = -1; while (++i < n) {streams[i].lineStart(); }},
-    lineEnd: function() { var i = -1; while (++i < n) {streams[i].lineEnd(); }},
-    polygonStart: function() { var i = -1; while (++i < n) {streams[i].polygonStart(); }},
-    polygonEnd: function() { var i = -1; while (++i < n) {streams[i].polygonEnd(); }}
+    point: function (x, y) {
+      var i = -1;
+      while (++i < n) {
+        streams[i].point(x, y);
+      }
+    },
+    sphere: function () {
+      var i = -1;
+      while (++i < n) {
+        streams[i].sphere();
+      }
+    },
+    lineStart: function () {
+      var i = -1;
+      while (++i < n) {
+        streams[i].lineStart();
+      }
+    },
+    lineEnd: function () {
+      var i = -1;
+      while (++i < n) {
+        streams[i].lineEnd();
+      }
+    },
+    polygonStart: function () {
+      var i = -1;
+      while (++i < n) {
+        streams[i].polygonStart();
+      }
+    },
+    polygonEnd: function () {
+      var i = -1;
+      while (++i < n) {
+        streams[i].polygonEnd();
+      }
+    },
   };
 }
 
 // A composite projection for Ecuador, configured by default for 960×500.
-export default function() {
+export default function () {
   var cache,
-      cacheStream,
+    cacheStream,
+    mainland = mercator().rotate([80, 1.5]),
+    mainlandPoint,
+    galapagos = mercator().rotate([90.73, 1]),
+    galapagosPoint,
+    point,
+    pointStream = {
+      point: function (x, y) {
+        point = [x, y];
+      },
+    };
 
-      mainland = mercator().rotate([80, 1.5]), mainlandPoint,
-      galapagos = mercator().rotate([90.73, 1]), galapagosPoint,
-
-      point, pointStream = {point: function(x, y) { point = [x, y]; }};
-
-      /*
+  /*
       var mainlandBbox = [[-81.5, 2.7], [-70.0, -6.0]];
       var galapagosBbox = [[-92.2, 0.58], [-88.8, -1.8]];
       */
 
   function mercatorEcuador(coordinates) {
-    var x = coordinates[0], y = coordinates[1];
-    return point = null,
-        (mainlandPoint.point(x, y), point) ||
-        (galapagosPoint.point(x, y), point);
+    var x = coordinates[0],
+      y = coordinates[1];
+    return (
+      (point = null),
+      (mainlandPoint.point(x, y), point) || (galapagosPoint.point(x, y), point)
+    );
   }
 
-  mercatorEcuador.invert = function(coordinates) {
+  mercatorEcuador.invert = function (coordinates) {
     var k = mainland.scale(),
-        t = mainland.translate(),
-        x = (coordinates[0] - t[0]) / k,
-        y = (coordinates[1] - t[1]) / k;
-        /*
+      t = mainland.translate(),
+      x = (coordinates[0] - t[0]) / k,
+      y = (coordinates[1] - t[1]) / k;
+    /*
         //How are the return values calculated:
         var c0 = galapagos(galapagosBbox[0]);
         var x0 = (c0[0] - t[0]) / k;
@@ -60,31 +96,47 @@ export default function() {
 
         console.info("p1 galapagos", x1 + ' - ' + y1);
         */
-        return (y >= -0.0676 && y< -0.026 && x >= -0.0857 && x < -0.0263 ? galapagos
-            : mainland).invert(coordinates);
+    return (
+      y >= -0.0676 && y < -0.026 && x >= -0.0857 && x < -0.0263
+        ? galapagos
+        : mainland
+    ).invert(coordinates);
   };
 
-  mercatorEcuador.stream = function(stream) {
-    return cache && cacheStream === stream ? cache : cache = multiplex([mainland.stream(cacheStream = stream), galapagos.stream(stream)]);
+  mercatorEcuador.stream = function (stream) {
+    return cache && cacheStream === stream
+      ? cache
+      : (cache = multiplex([
+          mainland.stream((cacheStream = stream)),
+          galapagos.stream(stream),
+        ]));
   };
 
-  mercatorEcuador.precision = function(_) {
-    if (!arguments.length) {return mainland.precision();}
+  mercatorEcuador.precision = function (_) {
+    if (!arguments.length) {
+      return mainland.precision();
+    }
     mainland.precision(_);
     galapagos.precision(_);
     return reset();
   };
 
-  mercatorEcuador.scale = function(_) {
-    if (!arguments.length) {return mainland.scale();}
+  mercatorEcuador.scale = function (_) {
+    if (!arguments.length) {
+      return mainland.scale();
+    }
     mainland.scale(_);
     galapagos.scale(_);
     return mercatorEcuador.translate(mainland.translate());
   };
 
-  mercatorEcuador.translate = function(_) {
-    if (!arguments.length) {return mainland.translate();}
-    var k = mainland.scale(), x = +_[0], y = +_[1];
+  mercatorEcuador.translate = function (_) {
+    if (!arguments.length) {
+      return mainland.translate();
+    }
+    var k = mainland.scale(),
+      x = +_[0],
+      y = +_[1];
     /*
     var c0 = mainland(mainlandBbox[0]);
    var x0 = (x - c0[0]) / k;
@@ -125,23 +177,29 @@ export default function() {
     ' * k - epsilon]])');*/
 
     mainlandPoint = mainland
-        .translate(_)
-        .clipExtent([[x - 0.0262 * k, y - 0.0734 * k],[x + 0.1741 * k, y + 0.079 * k]])
-        .stream(pointStream);
+      .translate(_)
+      .clipExtent([
+        [x - 0.0262 * k, y - 0.0734 * k],
+        [x + 0.1741 * k, y + 0.079 * k],
+      ])
+      .stream(pointStream);
 
     galapagosPoint = galapagos
-        .translate([x - 0.06 * k, y - 0.04 * k])
-        .clipExtent([[x - 0.0857 * k + epsilon, y - 0.0676 * k + epsilon],[x - 0.0263 * k - epsilon, y - 0.026 * k - epsilon]])
-        .stream(pointStream);
+      .translate([x - 0.06 * k, y - 0.04 * k])
+      .clipExtent([
+        [x - 0.0857 * k + epsilon, y - 0.0676 * k + epsilon],
+        [x - 0.0263 * k - epsilon, y - 0.026 * k - epsilon],
+      ])
+      .stream(pointStream);
 
     return reset();
   };
 
-  mercatorEcuador.fitExtent = function(extent, object) {
+  mercatorEcuador.fitExtent = function (extent, object) {
     return fitExtent(mercatorEcuador, extent, object);
   };
 
-  mercatorEcuador.fitSize = function(size, object) {
+  mercatorEcuador.fitSize = function (size, object) {
     return fitSize(mercatorEcuador, size, object);
   };
 
@@ -150,7 +208,7 @@ export default function() {
     return mercatorEcuador;
   }
 
-  mercatorEcuador.drawCompositionBorders = function(context) {
+  mercatorEcuador.drawCompositionBorders = function (context) {
     /*
     console.info("CLIP EXTENT: ", galapagos.clipExtent());
     console.info("UL BBOX:", mainland.invert([galapagos.clipExtent()[0][0], galapagos.clipExtent()[0][1]]));
@@ -169,9 +227,8 @@ export default function() {
     context.lineTo(ldgalapagos[0], ldgalapagos[1]);
     context.lineTo(llgalapagos[0], llgalapagos[1]);
     context.closePath();
-
   };
-  mercatorEcuador.getCompositionBorders = function() {
+  mercatorEcuador.getCompositionBorders = function () {
     var context = path();
     this.drawCompositionBorders(context);
     return context.toString();
